Rename ServerSideApis as ServerSideApi
