+++ src/store/sagas/localeSaga.ts
@@ -1,12 +1,12 @@
 import { all, put, call, take, takeEvery } from 'redux-saga/effects';
 
 import { ActionTypes } from '@actions/index';
-import * as ServerSideApis from '@src/apis/ServerSideApis';
+import * as ServerSideApi from '@src/apis/ServerSideApi';
 import { createAction } from '@actions/index';
 
 export function* getTranslate(action) {
   try {
-    const data = yield ServerSideApis.getTranslate(action.payload.locale);
+    const data = yield ServerSideApi.getTranslate(action.payload.locale);
     yield put(createAction(ActionTypes.GET_LOCALE_SUCCESS, {
       locale: data.locale
     }));
